Rename classes, refactor SenatorList
